Greatly optimized AI.getOccurrencesCount It now takes less than five seconds to train with blogstrain, used to takes more than five minutes (never waited until completion, so could be hours).

+++ test/mod6AI/ai/AITest.java
@@ -48,20 +48,20 @@ public class AITest {
 
     @Test
     public void testGetOccurrencesCount() throws Exception {
-        Map<String, Integer> occurrences = AI.getOccurrencesCount(
+        Map<String, Long> occurrences = AI.getOccurrencesCount(
                 Tokenizer.tokenize("I don't like Java.\nHowever Java 8 has some pretty nice features. :)"));
 
-        assertEquals(1, (int) occurrences.get("i"));
-        assertEquals(1, (int) occurrences.get(Tokenizer.removeAllNoneAlphanumeric("don't")));
-        assertEquals(1, (int) occurrences.get("like"));
-        assertEquals(2, (int) occurrences.get("java"));
-        assertEquals(1, (int) occurrences.get("however"));
-        assertEquals(1, (int) occurrences.get("8"));
-        assertEquals(1, (int) occurrences.get("has"));
-        assertEquals(1, (int) occurrences.get("some"));
-        assertEquals(1, (int) occurrences.get("pretty"));
-        assertEquals(1, (int) occurrences.get("nice"));
-        assertEquals(1, (int) occurrences.get("features"));
+        assertEquals(1, occurrences.get("i").longValue());
+        assertEquals(1, occurrences.get(Tokenizer.removeAllNoneAlphanumeric("don't")).longValue());
+        assertEquals(1, occurrences.get("like").longValue());
+        assertEquals(2, occurrences.get("java").longValue());
+        assertEquals(1, occurrences.get("however").longValue());
+        assertEquals(1, occurrences.get("8").longValue());
+        assertEquals(1, occurrences.get("has").longValue());
+        assertEquals(1, occurrences.get("some").longValue());
+        assertEquals(1, occurrences.get("pretty").longValue());
+        assertEquals(1, occurrences.get("nice").longValue());
+        assertEquals(1, occurrences.get("features").longValue());
         assertEquals(11, occurrences.size());
     }
 }
